refactor(tests): split ButtonGroupsBuilder.test.ts into 2 focused test files
Split 1365 LOC file into:
- ButtonGroupsBuilder.fixtures.ts - shared test setup and context
- ButtonGroupsBuilder.build.test.ts - build method tests (28 tests)
- ButtonGroupsBuilder.handlers.test.ts - onClick handler tests (10 tests)

Part of Issue #474: Split oversized test files
